fix reward type

+++ src/deeprl_hw2/core.py
@@ -211,7 +211,7 @@ class ReplayMemory:
         self.memory_size = args.replay_memory_size
         self.history_length = args.num_frames
         self.actions = np.zeros(self.memory_size, dtype = np.int8)
-        self.rewards = np.zeros(self.memory_size, dtype = np.int8)
+        self.rewards = np.zeros(self.memory_size)
         self.screens = np.zeros((self.memory_size, args.frame_height, args.frame_width), dtype = np.uint8)
         self.terminals = np.zeros(self.memory_size, dtype = np.bool)
         self.current = 0
